Refactor auth module into onion
