+++ backend/libs/auth/src/application.rs
@@ -1,0 +1,1 @@
+pub mod auth_use_cases;
