+++ backend/libs/auth/src/domain.rs
@@ -1,0 +1,3 @@
+pub mod jwt_secret;
+pub mod login_request;
+pub mod login_service;
