+++ backend/libs/auth/src/domain/principal.rs
@@ -1,0 +1,2 @@
+#[async_trait]
+pub trait Principal: Sync + Send {}
