+++ backend/libs/auth/src/infrastructure/auth_state.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use crate::auth_use_cases::AuthUseCases;
+use crate::application::auth_use_cases::AuthUseCases;
 
 #[derive(Clone)]
 pub struct AuthState(pub Arc<AuthUseCases>);
