+++ backend/libs/auth/src/infrastructure/auth_user_entity.rs
