+++ backend/libs/auth/src/router.rs
@@ -5,9 +5,10 @@ use axum::{
 };
 
 use crate::{
-    auth_state::AuthState,
-    login_handler::login,
-    test_protected_endpoint_handler::test_protected_endpoint,
+    AuthState,
+    infrastructure::{
+        login_handler::login, test_protected_endpoint_handler::test_protected_endpoint,
+    },
 };
 
 pub fn auth_router<S>() -> Router<S>
